Add conditional compilation over delta-table to fix a windows cross compilation issue
