+++ lib/customer-deposit-wallet/src/Cardano/Wallet/Deposit/IO/DB/Stub.hs
@@ -1,0 +1,2 @@
+module Cardano.Wallet.Deposit.IO.DB.Stub ()
+where
